Fix subtle bug
Bug fiexd short description: sometimes the gene cluster picture didn't get loaded. The reson is related to the Observable behavior.  The target issued element in neighbourGenes$ observable in NeighborGenesView is the second one. But sometimes there is only one value in the observable and that's the needed value. Also when searching the genes, the observable tipically gets fired more than twise, while when browsing in the gene cluster or reloading - twise at most. The code below is a sly and efficient way to deal with this.